upload fil

+++ src/app/components/views/admin/file-cursus/file-cursus/file-cursus.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { NgToastService } from 'ng-angular-popup';
 import { UploadFileService } from 'src/app/services/upload-file.service';
 
 @Component({
@@ -8,7 +9,7 @@ import { UploadFileService } from 'src/app/services/upload-file.service';
 })
 export class FileCursusComponent implements OnInit {
   ListFiles: any[] = [];
-  constructor(private documentService: UploadFileService) {}
+  constructor(private documentService: UploadFileService,private toast :NgToastService) {}
 
   ngOnInit(): void {
     this.documentService.getAllDocument().subscribe((result) => {
@@ -17,10 +18,16 @@ export class FileCursusComponent implements OnInit {
     });
   }
   delete(item: any) {
+    if (confirm('Voulez vous débloquer ce document?')) {
     this.documentService.deleteDocument(item.id).subscribe((result) => {
       let index = this.ListFiles.indexOf(item);
       this.ListFiles.splice(index, 1);
+      this.toast.error({
+        detail: ' Message',
+        summary: 'document est Supprimé',
+        duration: 2000,
+      })
       console.log(result);
-    });
+    });}
   }
 }
